backend/requirements.txt
Flask==3.0.2
requests==2.31.0
Werkzeug==3.0.2
Flask-Cors==4.0.0
gunicorn==20.1.0
google-cloud-pubsub